Adjust module layout for reader and writer modules
Rather than putting the main implementations of the reader and writer
into __init__.py for their relevant subpackages, move that out into
private "main" modules, and strip `__init__.py` down to just do imports.
