+++ src/mddj/api/reader/_errors.py
@@ -1,0 +1,6 @@
+class MissingRequiredField(LookupError):
+    pass
+
+
+class NotALowerBound(ValueError):
+    pass
